Teamnado 5120 - set release ver (#248)
* Add workload selection, success page, and name validation

This adds workload selection to automatically add applicable additional
repositories for RHEL EUS versions. These are set by name and version,
which maps to an array of repos all provided by the rhsm-api endpoint
`v2/products/RHEL/extended-update-support-versions` endpoint.

It also adds alphanumeric, -, and _ validation for the name of an
activation key.

It also adds a success page with a link to the key, and the ability to
close and continue looking at the list of keys.

* Add release version to activation key

This adds the release version to an activation key created with an eus
version. It also adds a tooltip when no eus sub exists