Increase clarity in create activation key wizard (#267)
This adds a couple small wording changes to the "Select Workload" page
of the create activation key wizard. It also adds a couple tooltips to
give extra information on hover.

While I was in the file, I noticed that we were checking string equality
in a couple places to determine which workload we were looking at, so I
moved those to index checks. This allows us to not have to worry about
changing strings when the text there changes.